Add ability to bulk publish pages into CaaS
