+++ tools/send-to-caas/bulksend.html
@@ -1,0 +1,29 @@
+<html>
+  <body>
+    <div>
+      <label for="excelfile">Excel File Url:</label>
+      <input id="excelfile" name="excelfile" type="text">
+    </div>
+    <div class="caas-env">
+      <label for="caas-env-select">CaaS Env</label>
+      <select name="1A" id="caas-env-select">
+        <option>Prod</option>
+        <option>Stage</option>
+        <option>Dev</option>
+      </select>
+    </div>
+    <div id="caas-draft-cb">
+      <input type="checkbox" id="draftcb" name="draftcb">
+      <label for="draftcb">Publish to Draft only</label>
+    </div>
+    <button id="bulkpublish" href="#">Bulk Publish</button>
+    <label for="errors">Errors:</label>
+    <textarea id="errors" name="errors" rows="10" cols="50"></textarea>
+    <label for="success">Success:</label>
+    <textarea id="success" name="success" rows="10" cols="50"></textarea>
+    <script type="module">
+      import init from './bulk-send-to-caas.js';
+      init();
+    </script>
+  </body>
+</html>
